requirements.txt
albumentations==0.4.3
cycler==0.10.0
decorator==4.4.1
dill==0.3.1.1
imageio==2.6.1
imgaug==0.2.6
joblib==0.14.0
jsonpickle==1.2
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.4
numexpr==2.7.0
numpy==1.17.4
opencv-python==4.1.1.26
openslide-python==1.1.1
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.0
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.2
seaborn==0.9.0
six==1.13.0
tables==3.6.1
tensorboardX==1.9
torch==1.1.0
torchvision==0.3.0
tqdm==4.38.0
umap==0.1.1
